refactor(lisp): split SpecValidator into Parser + execution facade (#899)
`lib/ptc_runner/lisp/spec_validator.ex` had grown to 965 lines mixing
two distinct concerns:

1. **Markdown extraction** — parsing the PTC-Lisp spec markdown for
   `; =>` example markers, handling single-line and multi-line forms,
   balancing parentheses across line breaks, parsing TODO/BUG markers,
   parsing expected values (literals, strings, keywords, collections,
   maps, vars), computing per-section content hashes. Pure string
   munging, no execution.

2. **Example execution + reporting** — loading the spec file, running
   each extracted example through `PtcRunner.Lisp.run/2`, aggregating
   pass/fail stats per section, handling negative tests.

The two halves never share state, so each could be touched without
re-reading the other's 500 lines — once they were separated.

## Structure

* New module `PtcRunner.Lisp.SpecValidator.Parser`
  (`lib/ptc_runner/lisp/spec_validator/parser.ex`, 697 lines) — public
  API: `extract_examples/1` and `extract_section_hashes/1`. All
  helpers (paren balance, backward scan, fragment detection, expected
  value parsing) are private to this module.
* `PtcRunner.Lisp.SpecValidator` slimmed to 366 lines, kept as the
  thin orchestration facade. Its public API is unchanged — callers
  (`Mix.Tasks.Ptc.ValidateSpec`, `Mix.Tasks.Ptc.UpdateSpecChecksums`,
  the test suite, doctests, hexdocs) need no edits.
* `extract_examples/1` (content variant) now uses `defdelegate` to
  Parser, so the doctest still resolves through `SpecValidator`.

## Verification

* `mix compile --warnings-as-errors`, `mix format`, `mix credo
  --strict` (via pre-commit hook) clean
* `mix test test/ptc_runner/lisp/spec_validator_test.exs` →
  49 tests, 0 failures (2 excluded — Babashka conformance)
* Full root suite: 4927 tests, 0 failures (379 doctests + 3
  properties + tests, 364 excluded)
* `mix ptc.validate_spec` runs end-to-end against the live spec:
  **392/392 examples pass** (identical to pre-split baseline) across
  all 12 spec sections

Closes #899
